requirements.txt
openai-whisper
yt-dlp
streamlit
langchain
langchain-groq
groq
python-dotenv
faiss-cpu
tiktoken
ffmpeg-python
imageio[ffmpeg]